### 2020-02-21 #2108 @kyleniemeyer
@openjournals/dev any ideas why the paper won't compile? I see there is a figure, but can't decipher the error message.
