### 2020-02-21 #2108 @arfon
:wave: @danhey - does this paper accompany a publication in a AAS journal?
