### 2020-02-22 #2108 @danhey
>  @danhey - does this paper accompany a publication in a AAS journal?

Yes! The paper is currently in review. I submitted under the dual-review system, so the editors are aware of the JOSS review
