### 2020-02-25 #2108 @arfon
@mbobra - I've just added some docs on how to handle these papers that are being jointly published with AAS: https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing

/ cc @crawfordsm
