### 2020-02-27 7296acb
fixing astropy imports
