+++ maelstrom/eddy.py
@@ -20,7 +20,7 @@ from pymc3.model import Model
 
 import exoplanet as xo
 from exoplanet.orbits import get_true_anomaly
-from astropy.stats import LombScargle
+from astropy.timeseries import LombScargle
 import tqdm
 
 __all__ = ["Eddy"]
