### 2020-02-28 #2125 @whedon
[REVIEW]: Maelstrom: A Python package for identifying companions to pulsating stars from their light travel time variations
**Submitting author:** @danhey (<a href="http://orcid.org/0000-0003-3244-5357">Daniel Hey</a>)
**Repository:** <a href="https://github.com/danhey/maelstrom" target ="_blank">https://github.com/danhey/maelstrom</a>
**Version:** v0.3
