### 2020-03-06 #2125 @arfon
:wave: @mbobra - the accompanying AAS paper is accepted (not yet published) and the DOI is `10.3847/1538-3881/ab7d38`. When you get a chance @danhey, please add this information to your `paper.md` YAML header as [described in our docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain):

```
# Optional fields if submitting to a AAS journal too, see this blog post:
# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
aas-journal: Astrophysical Journal <- The name of the AAS journal.
```
