+++ paper/paper.md
@@ -41,8 +41,8 @@ bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
+aas-doi: 10.3847/1538-3881/ab7d38
+aas-journal: Astronomical Journal
 ---
 
 # Summary
