### 2020-05-04 #2125 @danhey
Hi all, just a heads up that the paper has now been published -- https://doi.org/10.3847/1538-3881/ab7d38
