### 2020-05-29 #2125 @hpparvi
Dear all,

I find the code clear and well-documented. Installation works without issues, the tutorials in the user guide are clear (some could do with a bit cleaning up, but nothing major), and the real-life use-case notebooks reproducing the analyses reported in the associated AJ paper are also very useful.

The only two minor issues I can find with this submission is the lack of community guidelines (e.g. how to contribute to the code.) and docstring formatting. Community guidelines are required by the JOSS policy but should be a very quick thing to fix.

The second issue is that some of the docstrings use Google style while others use Numpy style. Combining both styles in a single file may confuse Sphinx, and some of the API documentation has problems with formatting (such as the documentation of the main Maelstrom class). However, this is mainly cosmetics (the docs are understandable) and not an acceptance-blocker.

So, I believe the code will be useful to the astrophysics community, and I recommend accepting the submission after the community guidelines (at least a "contributing" document) are added.
