### 2020-06-03 #2125 @hpparvi
@danhey has added the contribution document to the repository and I'm now happy to recommend accepting his contribution.
