### 2020-06-24 #2125 @mbobra
@danhey Thank you, we're almost there! I have a couple more small requests:
*  Could you please cut a new release and deposit it in Zenodo? (I see the most recent one is from February 16, which doesn't include the `CONTRIBUTING.md` file). And then re-update the author list?
* Also do you mind changing the DOI for the Murphy+2018 reference from the short link to https://doi.org/10.1093/mnras/stx3049?
