### 2020-06-30 439485e
Update DOI for JOSS paper
