+++ paper/paper.bib
@@ -80,7 +80,7 @@
   volume = {474},
   pages = {4322--4346},
   issn = {0035-8711},
-  doi = {10/gdk3dm},
+  doi = {https://doi.org/10.1093/mnras/stx3049},
   abstract = {Abstract.  The orbital parameters of binaries at intermediate periods (102\textendash{}103\,d) are difficult to measure with conventional methods and are very incomplete. We},
   journal = {Monthly Notices of the Royal Astronomical Society},
   language = {en},
