### 2020-06-30 #2125 @danhey
Hi @mbobra, I've updated the DOI! I'm happy to make a new release for Zenodo but the AAS paper points to the v0.2 Zenodo release, which I don't think can be changed now.
