### 2020-06-30 #2125 @mbobra
@danhey Thank you! hmmm, interesting. I am not sure what to do here.

@openjournals/joss-eics, submitting author @danhey submitted a joint publication with The Astronomical Journal. The AAS publication cites Zenodo release v0.2 of the author's package, `Maelstrom`. The AAS review process is over, and the paper has been accepted (and cannot be changed). However, the JOSS review is ongoing. JOSS referees requested changes to `Maelstrom` that the submitting author implemented. Cutting a new release with these changes would result in v0.3, thereby causing a mismatch with The Astronomical Journal paper. How do you recommend we proceed?
