### 2020-06-30 #2125 @arfon
> Cutting a new release with these changes would result in v0.3, thereby causing a mismatch with The Astronomical Journal paper. How do you recommend we proceed?

I think this mismatch is not ideal but is better than not having a new release with the changes from this review. So please create a `v0.3` release and post the DOI here.

/ cc @crawfordsm from the AAS-side of things.
